+++ disco/info.go
@@ -12,7 +12,6 @@ import (
 	"mellium.im/xmpp"
 	"mellium.im/xmpp/disco/info"
 	"mellium.im/xmpp/form"
-	"mellium.im/xmpp/internal/ns"
 	"mellium.im/xmpp/jid"
 	"mellium.im/xmpp/stanza"
 )
@@ -41,52 +40,12 @@ func (q InfoQuery) WriteXML(w xmlstream.TokenWriter) (int, error) {
 	return xmlstream.Copy(w, q.TokenReader())
 }
 
-// Identity is the type and category of a node on the network.
-// Normally one of the pre-defined Identity types should be used.
-type Identity struct {
-	XMLName  xml.Name `xml:"http://jabber.org/protocol/disco#info identity"`
-	Category string   `xml:"category,attr"`
-	Type     string   `xml:"type,attr"`
-	Name     string   `xml:"name,attr,omitempty"`
-	Lang     string   `xml:"http://www.w3.org/XML/1998/namespace lang,attr,omitempty"`
-}
-
-// TokenReader implements xmlstream.Marshaler.
-func (i Identity) TokenReader() xml.TokenReader {
-	start := xml.StartElement{
-		Name: xml.Name{Space: NSInfo, Local: "query"},
-		Attr: []xml.Attr{{
-			Name:  xml.Name{Local: "category"},
-			Value: i.Category,
-		}, {
-			Name:  xml.Name{Local: "type"},
-			Value: i.Type,
-		}},
-	}
-	if i.Name != "" {
-		start.Attr = append(start.Attr, xml.Attr{
-			Name: xml.Name{Local: "name"}, Value: i.Name,
-		})
-	}
-	if i.Lang != "" {
-		start.Attr = append(start.Attr, xml.Attr{
-			Name: xml.Name{Space: ns.XML, Local: "lang"}, Value: i.Lang,
-		})
-	}
-	return xmlstream.Wrap(nil, start)
-}
-
-// WriteXML implements xmlstream.WriterTo.
-func (i Identity) WriteXML(w xmlstream.TokenWriter) (int, error) {
-	return xmlstream.Copy(w, i.TokenReader())
-}
-
 // Info is a response to a disco info query.
 type Info struct {
 	InfoQuery
-	Identity []Identity     `xml:"identity"`
-	Features []info.Feature `xml:"feature"`
-	Form     []form.Data    `xml:"jabber:x:data x,omitempty"`
+	Identity []info.Identity `xml:"identity"`
+	Features []info.Feature  `xml:"feature"`
+	Form     []form.Data     `xml:"jabber:x:data x,omitempty"`
 }
 
 // TokenReader implements xmlstream.Marshaler.