+++ disco/info_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"mellium.im/xmpp/disco"
+	"mellium.im/xmpp/disco/info"
 )
 
 func TestMarshalQuery(t *testing.T) {
@@ -42,7 +43,7 @@ func TestMarshalQuery(t *testing.T) {
 }
 
 func TestUnmarshal(t *testing.T) {
-	infoResp := `<query node="test" xmlns='http://jabber.org/protocol/disco#info'>
+	const infoXML = `<query node="test" xmlns='http://jabber.org/protocol/disco#info'>
   <identity
       category='conference'
       type='text'
@@ -64,41 +65,41 @@ func TestUnmarshal(t *testing.T) {
     </field>
   </x>
 </query>`
-	var info disco.Info
-	err := xml.Unmarshal([]byte(infoResp), &info)
+	var infoResp disco.Info
+	err := xml.Unmarshal([]byte(infoXML), &infoResp)
 	if err != nil {
 		t.Fatalf("unexpected error unmarshaling: %v", err)
 	}
-	if info.Node != "test" {
-		t.Errorf("node did not unmarshal correctly: want=test, got=%s", info.Node)
+	if infoResp.Node != "test" {
+		t.Errorf("node did not unmarshal correctly: want=test, got=%s", infoResp.Node)
 	}
-	if l := len(info.Identity); l != 2 {
+	if l := len(infoResp.Identity); l != 2 {
 		t.Errorf("wrong number of identities: want=2, got=%d", l)
 	}
-	ident := disco.Identity{
+	ident := info.Identity{
 		XMLName:  xml.Name{Space: disco.NSInfo, Local: "identity"},
 		Category: "conference",
 		Type:     "text",
 		Name:     "Play-Specific Chatrooms",
 		Lang:     "en",
 	}
-	if ident != info.Identity[0] {
-		t.Errorf("wrong identity: want=%v, got=%v", ident, info.Identity[0])
+	if ident != infoResp.Identity[0] {
+		t.Errorf("wrong identity: want=%v, got=%v", ident, infoResp.Identity[0])
 	}
-	if l := len(info.Features); l != 3 {
+	if l := len(infoResp.Features); l != 3 {
 		t.Errorf("wrong number of features: want=3, got=%d", l)
 	}
-	if v := info.Features[0].Var; v != disco.NSInfo {
+	if v := infoResp.Features[0].Var; v != disco.NSInfo {
 		t.Errorf("wrong first feature: want=%s, got=%s", disco.NSInfo, v)
 	}
-	if info.Form == nil {
+	if infoResp.Form == nil {
 		t.Errorf("form was not unmarshaled")
 	}
 	const serverInfo = "http://jabber.org/network/serverinfo"
-	if s, ok := info.Form[0].GetString("FORM_TYPE"); !ok || s != serverInfo {
+	if s, ok := infoResp.Form[0].GetString("FORM_TYPE"); !ok || s != serverInfo {
 		t.Errorf("wrong value for FORM_TYPE: want=%s, got=%s", serverInfo, s)
 	}
-	if s, ok := info.Form[0].GetString("c2s_port"); !ok || s != "5222" {
+	if s, ok := infoResp.Form[0].GetString("c2s_port"); !ok || s != "5222" {
 		t.Errorf("wrong value for FORM_TYPE: want=5222, got=%s", s)
 	}
 }