+++ CHANGELOG.md
@@ -10,8 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking
 
-- disco: the `Feature` and `Item` types have been moved to the `info` and
-  `items` packages
+- disco: the `Feature`, `Identity`, and `Item` types have been moved to the
+  `info` and `items` packages
 - roster: rename `version` attribute to `ver`
 - roster: the `Push` callback now takes the roster version
 - roster: `FetchIQ` now takes a `roster.IQ` instead of a `stanza.IQ` so that the